Move CFStartAction + StartCFAction to cf module
