+++ src/frontend/packages/cloud-foundry/cf-types.ts
@@ -5,7 +5,5 @@ export const CF_ENDPOINT_TYPE = 'cf';
 export class CFEntityConfig implements EntityCatalogueEntityConfig {
   public endpointType = CF_ENDPOINT_TYPE;
 
-  constructor(public entityType: string) {
-
-  }
+  constructor(public entityType: string) { }
 }
