+++ src/frontend/packages/cloud-foundry/src/actions/app-event.actions.ts
@@ -1,8 +1,8 @@
 import { RequestOptions, URLSearchParams } from '@angular/http';
 
-import { appEventEntityType, cfEntityFactory } from '../cf-entity-factory';
 import { PaginatedAction, QParam } from '../../../store/src/types/pagination.types';
-import { CFStartAction } from '../../../store/src/types/request.types';
+import { appEventEntityType, cfEntityFactory } from '../cf-entity-factory';
+import { CFStartAction } from './cf-action.types';
 
 export const AppGetAllEvents = {
   GET_ALL: '[Application Event] Get all',
