+++ src/frontend/packages/cloud-foundry/src/actions/application-service-routes.actions.ts
@@ -1,5 +1,13 @@
 import { RequestOptions, URLSearchParams } from '@angular/http';
 
+import { getActions } from '../../../store/src/actions/action.helper';
+import {
+  createEntityRelationKey,
+  createEntityRelationPaginationKey,
+  EntityInlineChildAction,
+  EntityInlineParentAction,
+} from '../../../store/src/helpers/entity-relations/entity-relations.types';
+import { ICFAction } from '../../../store/src/types/request.types';
 import { CFEntityConfig } from '../../cf-types';
 import {
   applicationEntityType,
@@ -9,14 +17,7 @@ import {
   serviceBindingEntityType,
   serviceInstancesEntityType,
 } from '../cf-entity-factory';
-import {
-  createEntityRelationKey,
-  createEntityRelationPaginationKey,
-  EntityInlineChildAction,
-  EntityInlineParentAction,
-} from '../../../store/src/helpers/entity-relations/entity-relations.types';
-import { CFStartAction, ICFAction } from '../../../store/src/types/request.types';
-import { getActions } from '../../../store/src/actions/action.helper';
+import { CFStartAction } from './cf-action.types';
 
 export const ASSIGN_ROUTE = '[Application] Assign route';
 export const ASSIGN_ROUTE_SUCCESS = '[Application] Assign route success';
