+++ src/frontend/packages/cloud-foundry/src/actions/feature-flags.actions.ts
@@ -1,9 +1,10 @@
 import { RequestOptions, URLSearchParams } from '@angular/http';
 
-import { cfEntityFactory, featureFlagEntityType } from '../cf-entity-factory';
-import { PaginatedAction } from '../../../store/src/types/pagination.types';
-import { CFStartAction, RequestEntityLocation } from '../../../store/src/types/request.types';
 import { getActions } from '../../../store/src/actions/action.helper';
+import { PaginatedAction } from '../../../store/src/types/pagination.types';
+import { RequestEntityLocation } from '../../../store/src/types/request.types';
+import { cfEntityFactory, featureFlagEntityType } from '../cf-entity-factory';
+import { CFStartAction } from './cf-action.types';
 
 export class GetAllFeatureFlags extends CFStartAction implements PaginatedAction {
   constructor(public endpointGuid: string, public paginationKey: string) {
