+++ src/frontend/packages/cloud-foundry/src/actions/quota-definitions.actions.ts
@@ -1,5 +1,8 @@
 import { RequestMethod, RequestOptions, URLSearchParams } from '@angular/http';
 
+import { EntityInlineChildAction } from '../../../store/src/helpers/entity-relations/entity-relations.types';
+import { PaginatedAction } from '../../../store/src/types/pagination.types';
+import { ICFAction } from '../../../store/src/types/request.types';
 import { CFEntityConfig } from '../../cf-types';
 import {
   cfEntityFactory,
@@ -7,9 +10,7 @@ import {
   quotaDefinitionEntityType,
   spaceQuotaEntityType,
 } from '../cf-entity-factory';
-import { EntityInlineChildAction } from '../../../store/src/helpers/entity-relations/entity-relations.types';
-import { PaginatedAction } from '../../../store/src/types/pagination.types';
-import { CFStartAction, ICFAction } from '../../../store/src/types/request.types';
+import { CFStartAction } from './cf-action.types';
 
 export const GET_QUOTA_DEFINITIONS = '[QuotaDefinitions] Get all';
 export const GET_QUOTA_DEFINITIONS_SUCCESS = '[QuotaDefinitions] Get all success';
