+++ src/frontend/packages/cloud-foundry/src/actions/security-groups-actions.ts
@@ -1,10 +1,13 @@
 import { RequestOptions } from '@angular/http';
 
-import { cfEntityFactory, securityGroupEntityType, spaceEntityType } from '../cf-entity-factory';
-import { createEntityRelationKey, EntityInlineParentAction } from '../../../store/src/helpers/entity-relations/entity-relations.types';
-import { PaginatedAction } from '../../../store/src/types/pagination.types';
-import { CFStartAction } from '../../../store/src/types/request.types';
 import { getActions } from '../../../store/src/actions/action.helper';
+import {
+  createEntityRelationKey,
+  EntityInlineParentAction,
+} from '../../../store/src/helpers/entity-relations/entity-relations.types';
+import { PaginatedAction } from '../../../store/src/types/pagination.types';
+import { cfEntityFactory, securityGroupEntityType, spaceEntityType } from '../cf-entity-factory';
+import { CFStartAction } from './cf-action.types';
 
 export class GetAllSecurityGroups extends CFStartAction implements PaginatedAction, EntityInlineParentAction {
   constructor(
