+++ src/frontend/packages/cloud-foundry/src/actions/service-bindings.actions.ts
@@ -1,10 +1,10 @@
 import { RequestOptions, URLSearchParams } from '@angular/http';
 
-import { cfEntityFactory, serviceBindingEntityType } from '../cf-entity-factory';
-import { PaginatedAction, PaginationParam } from '../../../store/src/types/pagination.types';
-import { CFStartAction, ICFAction } from '../../../store/src/types/request.types';
 import { getActions } from '../../../store/src/actions/action.helper';
-
+import { PaginatedAction, PaginationParam } from '../../../store/src/types/pagination.types';
+import { ICFAction } from '../../../store/src/types/request.types';
+import { cfEntityFactory, serviceBindingEntityType } from '../cf-entity-factory';
+import { CFStartAction } from './cf-action.types';
 
 export const DELETE_SERVICE_BINDING_ACTION = '[ Service Instances ] Delete Service Binding';
 export const DELETE_SERVICE_BINDING_ACTION_SUCCESS = '[ Service Instances ] Delete Service Binding success';
