+++ src/frontend/packages/cloud-foundry/src/actions/service-instances.actions.ts
@@ -1,5 +1,12 @@
 import { Headers, RequestOptions, URLSearchParams } from '@angular/http';
 
+import { getActions } from '../../../store/src/actions/action.helper';
+import {
+  createEntityRelationKey,
+  EntityInlineParentAction,
+} from '../../../store/src/helpers/entity-relations/entity-relations.types';
+import { PaginatedAction } from '../../../store/src/types/pagination.types';
+import { ICFAction } from '../../../store/src/types/request.types';
 import {
   applicationEntityType,
   cfEntityFactory,
@@ -12,10 +19,7 @@ import {
   servicePlanEntityType,
   spaceEntityType,
 } from '../cf-entity-factory';
-import { createEntityRelationKey, EntityInlineParentAction } from '../../../store/src/helpers/entity-relations/entity-relations.types';
-import { PaginatedAction } from '../../../store/src/types/pagination.types';
-import { CFStartAction, ICFAction } from '../../../store/src/types/request.types';
-import { getActions } from '../../../store/src/actions/action.helper';
+import { CFStartAction } from './cf-action.types';
 
 export const DELETE_SERVICE_BINDING = '[Service Instances] Delete service binding';
 export const UPDATE_SERVICE_INSTANCE_SUCCESS = getActions('Service Instances', 'Update Service Instance')[1];
