+++ src/frontend/packages/cloud-foundry/src/actions/service-plan.actions.ts
@@ -1,5 +1,8 @@
 import { RequestOptions, URLSearchParams } from '@angular/http';
 
+import { getActions } from '../../../store/src/actions/action.helper';
+import { createEntityRelationKey } from '../../../store/src/helpers/entity-relations/entity-relations.types';
+import { PaginatedAction } from '../../../store/src/types/pagination.types';
 import {
   applicationEntityType,
   cfEntityFactory,
@@ -10,10 +13,7 @@ import {
   servicePlanEntityType,
   spaceEntityType,
 } from '../cf-entity-factory';
-import { createEntityRelationKey } from '../../../store/src/helpers/entity-relations/entity-relations.types';
-import { PaginatedAction } from '../../../store/src/types/pagination.types';
-import { CFStartAction } from '../../../store/src/types/request.types';
-import { getActions } from '../../../store/src/actions/action.helper';
+import { CFStartAction } from './cf-action.types';
 
 export class GetServicePlanServiceInstances extends CFStartAction implements PaginatedAction {
   constructor(
